requirements-prod.txt
# Corporate Intelligence Platform - Production Dependencies
# Pinned versions for reproducible builds
# Generated: 2025-10-03
# Python: >=3.10

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
pydantic==2.5.2
pydantic-settings==2.1.0

# ASGI Server & Workers
uvloop==0.19.0
httptools==0.6.1

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.12.1
pgvector==0.2.4

# Data Processing
pandas==2.1.3
numpy==1.26.2
pandera==0.17.2
great-expectations==0.18.5
dbt-core==1.7.3
dbt-postgres==1.7.3

# Orchestration
prefect==2.14.9
prefect-dask==0.2.7
ray[default]==2.8.1
dask[complete]==2023.12.0

# Caching & Storage
redis==5.0.1
aiocache[redis]==0.12.2
hiredis==2.2.3
minio==7.2.0

# Observability & Monitoring
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-sqlalchemy==0.42b0
opentelemetry-exporter-otlp==1.21.0
prometheus-client==0.19.0
loguru==0.7.2
sentry-sdk[fastapi]==1.39.1

# Document Processing
pypdf==3.17.4
pdfplumber==0.10.3
python-docx==1.1.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Financial Data
yfinance==0.2.33
alpha-vantage==2.3.1
sec-edgar-api==1.0.1

# Visualization
plotly==5.18.0
dash==2.14.2
kaleido==0.2.1

# API & HTTP
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.7

# Data Validation
email-validator==2.1.0
python-dotenv==1.0.0
pyyaml==6.0.1

# Performance
orjson==3.9.10
ujson==5.9.0

# Testing (minimal for smoke tests)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
